test(pytest): guard all collected tests against head drift
Move the #2316 HEAD guard to the repository pytest root so both tests/ and test/ collections fail when a test leaves the real repo HEAD moved or unreadable.
