+++ tests/conftest.py
@@ -71,43 +71,3 @@ def temp_test_dir(tmp_path: Path) -> Path:
     test_dir.mkdir(parents=True, exist_ok=True)
     return test_dir
 
-
-# ---------------------------------------------------------------------------
-# #2316: guard against a git-using test mutating the real repo worktree HEAD.
-# A test that runs a git mutation (commit/reset/checkout) with a cwd that
-# resolves to this repo instead of an isolated tmp repo can move the checked-out
-# branch and leave the tree dirty -- dangerous for autofix agents. This fixture
-# turns that silent corruption into a deterministic, test-pinpointing failure.
-# ---------------------------------------------------------------------------
-
-import subprocess as _subprocess  # noqa: E402
-
-
-def _real_repo_head() -> str | None:
-    try:
-        out = _subprocess.run(
-            ["git", "rev-parse", "HEAD"],
-            cwd=str(PROJECT_ROOT),
-            capture_output=True,
-            text=True,
-            timeout=10,
-        )
-    except (OSError, _subprocess.SubprocessError):
-        return None
-    return out.stdout.strip() if out.returncode == 0 else None
-
-
-@pytest.fixture(autouse=True)
-def _guard_real_repo_head():
-    """Fail the test if it moved the real repo HEAD (#2316)."""
-    before = _real_repo_head()
-    yield
-    after = _real_repo_head()
-    if before and after and before != after:
-        pytest.fail(
-            f"#2316: this test mutated the REAL repo HEAD "
-            f"({before[:8]} -> {after[:8]}). A git mutation ran in the repo "
-            f"worktree instead of an isolated tmp repo. Isolate it: init a repo "
-            f"in tmp_path and run every git command with cwd=<tmp repo>.",
-            pytrace=False,
-        )
